aocpu: fix the source compilation warnings. [1/1]
PD#SWPL-43528

Problem:
need eliminate the compile warnings.

Solution:
fix the warnings.

Verify:
build pass.

+++ bl30/src_ao/demos/amlogic/n200/t7/interrupt_control.c
@@ -335,7 +335,7 @@ int eclic_map_interrupt(uint32_t ulIrq, uint32_t src)
 	return 0;
 }
 
-int eclic_interrupt_inner[SOC_ECLIC_NUM_INTERRUPTS] = {0};
+uint32_t eclic_interrupt_inner[SOC_ECLIC_NUM_INTERRUPTS] = {0};
 extern uint32_t vector_base;
 
 int RegisterIrq(uint32_t int_num, uint32_t int_priority, function_ptr_t handler) {
@@ -351,7 +351,7 @@ int RegisterIrq(uint32_t int_num, uint32_t int_priority, function_ptr_t handler)
 	}
 	eclic_interrupt_inner[irq - ECLIC_INTERNAL_NUM_INTERRUPTS] = int_num;
 
-	*(&vector_base + irq) = handler;
+	*(&vector_base + irq) = (uint32_t)handler;
 	eclic_set_irq_pri(irq, int_priority);
 
 	return 0;
